Issue #19: Reorganized routes and core modules initialization procedures.
