+++ botCommands/startCommand.py
@@ -4,7 +4,7 @@ import telebot  # pyTelegramBotAPI
 
 from core.helpers.time import getTimeStamp
 from core.logger import getLogger
-from core.appConfig import appConfig, PROJECT_INFO
+from core.appConfig import LOCAL, appConfig, PROJECT_INFO
 from core.utils import debugObj
 
 from botApp import botApp
@@ -21,27 +21,33 @@ def startCommand(chat: telebot.types.Chat):
     username = chat.username
     first_name = chat.first_name
     name = first_name if first_name else username
-    obj = {
+    debugItems = {
         'timeStr': getTimeStamp(),
         'chatId': chatId,
         'username': username,
         'first_name': first_name,
         'LOCAL': appConfig.get('LOCAL'),
     }
-    logContent = '\n'.join(
-        [
-            'startCommand',
-            debugObj(obj),
-        ]
-    )
-    content = '\n\n'.join(
-        [
-            'Hi, %s!' % name,
-            'Welcome to the TubeCaster bot!',
-            'The bot version is: %s.' % PROJECT_INFO,
-            'Use /cast to cast the youtube video to an audio. Type /help to find all commands.',
-        ]
-    )
+    logItems = [
+        'startCommand',
+        debugObj(debugItems),
+    ]
+    logContent = '\n'.join(logItems)
+    msgItems = [
+        'Hi, %s!' % name,
+        'Welcome to the TubeCaster bot!',
+        'The bot version is: %s.' % PROJECT_INFO,
+        'Use /cast to cast the youtube video to an audio. Type /help to find all commands.',
+        ' '.join(
+            [
+                'ATTENTION: Sorry, but at the moment the bot works in the LOCAL mode!',
+                " Probably, it won't be able to download any actual videos. Try a bit later please.",
+            ]
+        )
+        if LOCAL
+        else None,
+    ]
+    content = '\n\n'.join(msgItems)
     logger.info(logContent)
     # Show menu
     markup = createCommonButtonsMarkup()
