+++ botRoutes/_stopRoute.py
@@ -1,21 +1,12 @@
 # -*- coding:utf-8 -*-
 
-import time
-import traceback
 from flask import Response
-from flask import request
 
-from core.appConfig import appConfig, PROJECT_INFO
-from core.helpers.errors import errorToString
 from core.helpers.runtime import getModPath
 from core.helpers.time import getTimeStamp
-from core.logger import getLogger, loggerConfig
-from core.utils import debugObj
-from core.utils.generic import dictFromModule
+from core.logger import getLogger
 
 from botApp import botApp
-from botCore import botConfig
-from botCore.botConfig import WEBHOOK_URL
 
 from .botRoutes import botRoutes
 
@@ -24,9 +15,6 @@ startTimeStr = getTimeStamp()
 _logger = getLogger(getModPath())
 
 logTraceback = False
-
-
-# Real code start...
 
 
 @botRoutes.route('/stop')
@@ -38,7 +26,3 @@ def stopRoute():
     botApp.stop_bot()
     _logger.info('stopRoute')
     return Response('The webhook has been deleted', headers={'Content-type': 'text/plain'})
-
-
-# Module exports...
-__all__ = []
