+++ core/helpers/strings.py
@@ -1,0 +1,4 @@
+def truncStr(str: str, maxLen: int):
+    if len(str) >= maxLen - 3:
+        str = str[: maxLen - 3] + '...'
+    return str
